refactor(web): split templates into per-file modules with nicheId/category
Template gains nicheId (1:1 with NichePreset.id, null for datasets) and
category (industry | dataset). Graphs are unchanged; ids are immutable
because ?template=<id> deep links are public CTAs.
